Store min_age when adding new criteria
The criteria I am most interested in is the minimum age you have to be
to book a vaccination.

I'd like to pre-extract this in the JSON file so that I don't have to
extract it from the text elsewhere, such as in the @canigetjabbed bot.

Here I'm extracting the age when I add to the criteria store.