Fix storing duplicate criteria
Last week the NHS website was occasionally returning a cached version of
the page with out of date criteria (36+).

Because we were only checking against the 'latest' stored criteria (34+
at the time) we were not detecting the duplicate and ended up with three
36+ entries.

This change checks against all stored criteria for duplicates.

It is possible that published criteria may move 'backwards'
deliberately, eg the age range deliberately moves back to 36+. This
would be rejected by CriteriaStore#add, but if that happens I will deal
with it then.